requirements.txt
Flask
flask-cors
gunicorn
tensorflow
numpy
pillow
opencv-python-headless